user-experience-and-animations

+++ Views/TodoListView.swift
@@ -11,27 +11,36 @@ struct TodoListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+            else {
+                List{
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                }
+                .listStyle(PlainListStyle())
+            }
         }
-        .listStyle(PlainListStyle())
-        .navigationTitle("To-do List 📝")
+        .navigationTitle("To-do List  📝")
         .toolbar {
-            ToolbarItem(placement: .navigationBarLeading){
+            ToolbarItem(placement: .navigationBarLeading) {
                 EditButton()
             }
-            ToolbarItem(placement: .navigationBarTrailing){
-                NavigationLink(destination: AddView()) {
-                    Text("Add")
+            ToolbarItem(placement: .navigationBarTrailing) {
+                NavigationLink("Add") {
+                    AddView()
                 }
             }
         }
